requirements.txt
mysql-connector-python==9.1.0
Pillow==9.5.0
python-dotenv==1.0.1